requirements.txt
Django==3.1.5
djangorestframework==3.12.2
environs~=9.3.0
dj-database-url==0.5.0
gunicorn==20.0.4
psycopg2==2.8.6
django-filter~=2.4.0